Resolved CID 351470 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument this->_spellOverrideIdStore.size() to format specifier %u was expected to have type unsigned int but has type unsigned long

+++ src/world/Storage/MySQLDataStore.cpp
@@ -2646,7 +2646,7 @@ void MySQLDataStore::loadSpellOverrideTable()
 
     delete spelloverride_result;
 
-    LogDetail("MySQLDataLoads : %u spell overrides loaded.", _spellOverrideIdStore.size());
+    LogDetail("MySQLDataLoads : %u spell overrides loaded.", static_cast<uint32_t>(_spellOverrideIdStore.size()));
 }
 
 void MySQLDataStore::loadNpcGossipTextIdTable()
